Move FluidUnits into core.lib.fluids beside the fluid storage API
